Signed-signature to notes
Before, due to how notes were handled, any player can find a way to duplicate notes without the plugin caring if it's a duplicated copy.

This fix adds a unique signature to each note such that once a note is used, any other notes with the same signature will not work. This will prevent dupes.